// src/generated/kubb/zod/webSocketResponseCreateEventSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { createResponseBodySchema } from "./createResponseBodySchema.ts";
import { z } from "zod";

/**
 * @description A client-sent WebSocket event that starts a Responses turn. It uses the normal response-create body plus a required `type` field.
 */
export const webSocketResponseCreateEventSchema = z
  .lazy(() => createResponseBodySchema)
  .and(
    z.object({
      type: z
        .enum(["response.create"])
        .default("response.create")
        .describe("The client event type. Always `response.create`."),
      generate: z.optional(
        z
          .boolean()
          .describe(
            "Whether the server should generate model output for this turn. Set to `false` to warm up request state without producing model output.",
          ),
      ),
    }),
  )
  .describe(
    "A client-sent WebSocket event that starts a Responses turn. It uses the normal response-create body plus a required `type` field.",
  );